+++ packages/app/features/top/ChatListScreen.tsx
@@ -15,7 +15,6 @@ import { useChatSettingsNavigation } from '../../hooks/useChatSettingsNavigation
 import { useFilteredChats } from '../../hooks/useFilteredChats';
 import { TabName } from '../../hooks/useFilteredChats';
 import { useGroupActions } from '../../hooks/useGroupActions';
-import { startAgentGroupNavigationLockFailsafe } from '../../hooks/useAgentGroupOnboardingLock';
 import { useScrollToTabTop } from '../../hooks/useScrollToTabTop';
 import { useSyncStatus } from '../../hooks/useSyncStatus';
 import { reportChatListFirstPaint } from '../../lib/chatListSettleTelemetry';
@@ -41,10 +40,7 @@ import { ChatList, ChatListItemData } from '../chat-list/ChatList';
 import { ChatListSearch } from '../chat-list/ChatListSearch';
 import { ChatListTabs } from '../chat-list/ChatListTabs';
 import { CreateChatSheet, CreateChatSheetMethods } from './CreateChatSheet';
-import {
-  canClaimAgentOnboardingLanding,
-  claimAgentOnboardingLanding,
-} from './agentOnboardingLanding';
+import { useAgentOnboardingLandingConsumer } from './useAgentOnboardingLandingConsumer';
 import {
   getGroupInviteSheetState,
   isGroupInviteReady,
@@ -74,7 +70,7 @@ export function ChatListScreenView({
   previewGroupFromInviteNotification?: boolean;
   focusedChannelId?: string;
 }) {
-  const { navigation, navigateToGroup, navigateToChannel, resetToChannel } =
+  const { navigation, navigateToGroup, navigateToChannel } =
     useRootNavigation();
   const [personalInviteOpen, setPersonalInviteOpen] = useState(false);
   const personalInvite = db.personalInviteLink.useValue();
@@ -100,60 +96,7 @@ export function ChatListScreenView({
     enabled: isFocused,
   });
 
-  const onboardingLanding = db.agentOnboardingLanding.useValue();
-  const consumedOnboardingLanding = useRef(false);
-  const resetToChannelRef = useRef(resetToChannel);
-  resetToChannelRef.current = resetToChannel;
-  useEffect(() => {
-    if (
-      !canClaimAgentOnboardingLanding(onboardingLanding) ||
-      consumedOnboardingLanding.current
-    ) {
-      return;
-    }
-    let active = true;
-
-    void (async () => {
-      while (active && !consumedOnboardingLanding.current) {
-        try {
-          const channel = await db.getChannel({
-            id: onboardingLanding.channelId,
-          });
-          if (channel) {
-            // Furnishing may have taken much longer than the lock failsafe.
-            // Start its clock at the actual handoff so the setup chat gets the
-            // full bounded lock window once it becomes visible.
-            await startAgentGroupNavigationLockFailsafe(
-              onboardingLanding.groupId
-            );
-            // Claim this handoff durably before resetting navigation. The reset
-            // remounts ChatListScreen, so component-local state alone cannot
-            // prevent the new instance from consuming the same handoff again.
-            await db.agentOnboardingLanding.setValue(
-              claimAgentOnboardingLanding(onboardingLanding)
-            );
-            consumedOnboardingLanding.current = true;
-            resetToChannelRef.current(onboardingLanding.channelId, {
-              backToGroupIndex: true,
-              disableTransition: true,
-              groupId: onboardingLanding.groupId,
-            });
-            return;
-          }
-        } catch (error) {
-          logger.trackError('Failed to consume agent onboarding landing', {
-            error,
-            ...onboardingLanding,
-          });
-        }
-        await new Promise((resolve) => setTimeout(resolve, 500));
-      }
-    })();
-
-    return () => {
-      active = false;
-    };
-  }, [onboardingLanding]);
+  useAgentOnboardingLandingConsumer();
   const { performGroupAction } = useGroupActions();
 
   const handleInviteFriends = useCallback(() => {
